refactor: update redis configuration in docker-compose and implement useD3Graph hook for improved graph rendering in useStreamGraph
