+++ docker-compose.dev.yaml
@@ -34,7 +34,6 @@ services:
 
   redis:
     image: redis:latest
-    command: ["redis-server", "--requirepass", "root"]
     volumes:
       - redis-data:/data 
     ports:
@@ -43,6 +42,7 @@ services:
       - app-network
     environment:
       - REDIS_PASSWORD=root  # Configurando a senha
+    command: ["redis-server", "--requirepass", "root", "--appendonly", "yes", "--appendfsync", "always", "--save", "60 1000"]
 
 networks:
   app-network:
